pyproject.toml
[project]
name = "oclabs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.127.0",
    "openai-agents>=0.6.4",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
